Tutorial on using Guild AI to track TensorFlow experiments
Guild AI is an open source toolset for running and tracking
experiments. It's always had strong TensorFlow support. The link is to
a step-by-step tutorial that shows how the TensorFlow "get started"
sample code can be run without modification to generate runs, evaluate
results, and find hyperparameters that improve performance.